pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-qdrant>=0.2.1",
    "langgraph>=0.6.8",
    "pymupdf>=1.26.4",
    "pyowm>=3.5.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.4",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
]

[tool.uv.workspace]
members = [
    "personal",
]
